Prefences Modal: Add show weekend as an option
