--- static/js/redux/ui/containers/modals/preference_modal_container.jsx
@@ -1,29 +1,0 @@
-/*
-Copyright (C) 2017 Semester.ly Technologies, LLC
-
-Semester.ly is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Semester.ly is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-*/
-
-import { connect } from "react-redux";
-import PreferenceModal from "../../modals/preference_modal";
-import { preferencesActions } from "../../../state/slices/preferencesSlice";
-
-const mapStateToProps = (state) => ({
-  isVisible: state.preferences.isModalVisible,
-  withConflicts: state.preferences.try_with_conflicts,
-});
-
-const PreferenceModalContainer = connect(mapStateToProps, {
-  togglePreferenceModal: preferencesActions.togglePreferenceModal,
-  toggleConflicts: preferencesActions.toggleConflicts,
-})(PreferenceModal);
-
-export default PreferenceModalContainer;
